requirements.txt
pymongo>=4.6.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
tenacity>=8.2.0